Added the concept of an Immutable TK, which should not ever change - as I found I was changing a TK stored in the genbook tree.
However, it is possible to change it by assigning it to a module and then
moving the module around. However, this is forbidden behaviour.
When used in python, it will check to see whether it has changed to detect
this behaviour, and throw an error if it does.

Also, added a custom delimiter parameter to breadcrumb